+++ CHANGELOG.md
@@ -10,6 +10,7 @@ IMPROVEMENTS:
 
 - Differentiate external crypto service test errors when setting DPKI callbacks by separating into multiple error types.
 - Return HTTP response 503 with correct reason when waiting for DPKI callback URLs to be set (if configured to use external crypto service).
+- Change storage for AS service callback URLs from storing in cache DB to write to file in plain text for easier URL change when the process is not running.
 
 BUG FIXES: